+++ src/kde/config/ImageTypesTab.cpp
@@ -103,15 +103,13 @@ public:
 	/** Other ImageTypesTabPrivate functions **/
 
 	/**
-	 * Initialize strings.
+	 * Initialize the Credits label.
 	 */
-	void initStrings(void);
+	void initCreditsLabel(void);
 
 public:
-	// Last ComboBox added.
-	// Needed in order to set the correct
-	// tab order for the credits label.
-	QComboBox *cboImageType_lastAdded;
+	// Credits label
+	QLabel *lblCredits;
 
 	// Temporary QSettings object.
 	// Set and cleared by ImageTypesTab::save();
@@ -122,15 +120,15 @@ public:
 
 ImageTypesTabPrivate::ImageTypesTabPrivate(ImageTypesTab* q)
 	: q_ptr(q)
-	, cboImageType_lastAdded(nullptr)
+	, lblCredits(nullptr)
 	, pSettings(nullptr)
 { }
 
 ImageTypesTabPrivate::~ImageTypesTabPrivate()
 {
-	// cboImageType_lastAdded should be nullptr.
-	// (Cleared by finishComboBoxes().)
-	assert(cboImageType_lastAdded == nullptr);
+	// lblCredits should *not* be nullptr.
+	// (Created by initCreditsLabel().)
+	assert(lblCredits != nullptr);
 
 	// pSettings should be nullptr,
 	// since it's only used when saving.
@@ -214,12 +212,6 @@ void ImageTypesTabPrivate::createComboBox(unsigned int cbid)
 	cbo->setProperty("rp-config.cbid", cbid);
 	QObject::connect(cbo, SIGNAL(currentIndexChanged(int)),
 			 q, SLOT(cboImageType_currentIndexChanged()));
-
-	// Adjust the tab order.
-	if (cboImageType_lastAdded) {
-		q->setTabOrder(cboImageType_lastAdded, cbo);
-	}
-	cboImageType_lastAdded = cbo;
 }
 
 /**
@@ -258,15 +250,8 @@ void ImageTypesTabPrivate::addComboBoxStrings(unsigned int cbid, int max_prio)
  */
 void ImageTypesTabPrivate::finishComboBoxes(void)
 {
-	if (!cboImageType_lastAdded) {
-		// Nothing to do here.
-		return;
-	}
-
-	// Set the tab order for the credits label.
-	Q_Q(ImageTypesTab);
-	q->setTabOrder(cboImageType_lastAdded, ui.lblCredits);
-	cboImageType_lastAdded = nullptr;
+	// Nothing to do here.
+	return;
 }
 
 /**
@@ -349,10 +334,31 @@ void ImageTypesTabPrivate::cboImageType_setPriorityValue(unsigned int cbid, unsi
 /** Other ImageTypesTabPrivate functions **/
 
 /**
- * Initialize strings.
+ * Initialize the Credits label.
  */
-void ImageTypesTabPrivate::initStrings(void)
+void ImageTypesTabPrivate::initCreditsLabel(void)
 {
+	// Create the Credits label if it hasn't been created yet.
+	if (!lblCredits) {
+		RP_Q(ImageTypesTab);
+
+		QSizePolicy sizePolicy(QSizePolicy::Preferred, QSizePolicy::Maximum);
+		sizePolicy.setHorizontalStretch(0);
+		sizePolicy.setVerticalStretch(0);
+
+		lblCredits = new QLabel(q);
+		sizePolicy.setHeightForWidth(lblCredits->sizePolicy().hasHeightForWidth());
+		lblCredits->setSizePolicy(sizePolicy);
+		lblCredits->setFocusPolicy(Qt::StrongFocus);
+		lblCredits->setTextFormat(Qt::RichText);
+		lblCredits->setAlignment(Qt::AlignBottom|Qt::AlignLeading|Qt::AlignLeft);
+		lblCredits->setWordWrap(true);
+		lblCredits->setOpenExternalLinks(true);
+		lblCredits->setTextInteractionFlags(Qt::LinksAccessibleByKeyboard|Qt::LinksAccessibleByMouse);
+
+		ui.vboxMain->addWidget(lblCredits);
+	}
+
 	// tr: External image credits.
 	QString sCredits = QC_("ImageTypesTab",
 		"GameCube, Wii, Wii U, Nintendo DS, and Nintendo 3DS external images\n"
@@ -362,7 +368,7 @@ void ImageTypesTabPrivate::initStrings(void)
 
 	// Replace "\n" with "<br/>".
 	sCredits.replace(QChar(L'\n'), QLatin1String("<br/>"));
-	ui.lblCredits->setText(sCredits);
+	lblCredits->setText(sCredits);
 }
 
 /** ImageTypesTab **/
@@ -374,11 +380,14 @@ ImageTypesTab::ImageTypesTab(QWidget *parent)
 	Q_D(ImageTypesTab);
 	d->ui.setupUi(this);
 
-	// Initialize strings.
-	d->initStrings();
-
 	// Create the control grid.
 	d->createGrid();
+
+	// Create the credits label.
+	// NOTE: Creating it here instead of in the UI file in order to work around
+	// issues with tab ordering. (setTabOrder() isn't working if the control
+	// right before it is the final QComboBox that was created...)
+	d->initCreditsLabel();
 }
 
 ImageTypesTab::~ImageTypesTab()
@@ -396,7 +405,7 @@ void ImageTypesTab::changeEvent(QEvent *event)
 		// Retranslate the UI.
 		Q_D(ImageTypesTab);
 		d->ui.retranslateUi(this);
-		d->initStrings();
+		d->initCreditsLabel();
 	}
 
 	// Pass the event to the base class.